This commit adds support for nodes marked as Terminating by Gardener, ensuring the load balancer is updated immediately to avoid sending traffic to terminating nodes.
Kubernetes upstream recently optimized API requests to load balancer services, but this unintentionally ignored node status changes made by Gardener.
As a result, the load balancer was updated too late, leading to brief periods of packet loss.
To address this, the commit reverts that upstream behavior and explicitly ensures that UpdateLoadBalancer() is called when a node enters the Terminating state,
ensuring timely load balancer updates and bringing connection interrupts to zero.

Additional Note: 1.30,1.31 makes the featureGate as GA, therefore it is here removed from the code.

Signed-off-by: Niclas Schad <niclas.schad@stackit.cloud>